+++ CHANGELOG.md
@@ -174,6 +174,8 @@
   it robust against project name changes.
 - [Added a button to return from an opened project back to the project
   dashboard.][6474]
+- [Keyboard shortcuts for graph editing are now disabled when the full-screen
+  visualization is active.][6844]
 
 [6279]: https://github.com/enso-org/enso/pull/6279
 [6421]: https://github.com/enso-org/enso/pull/6421
@@ -184,6 +186,7 @@
 [6752]: https://github.com/enso-org/enso/pull/6752
 [6719]: https://github.com/enso-org/enso/pull/6719
 [6474]: https://github.com/enso-org/enso/pull/6474
+[6844]: https://github.com/enso-org/enso/pull/6844
 
 #### EnsoGL (rendering engine)